main_code/requirements.txt
flask
gunicorn==20.1.0
google-cloud-aiplatform
google-cloud
numpy